tests/wpt/tests/speculation-rules/prerender/restriction-service-worker-postmessage.https.html

<!DOCTYPE html>
<title>postMessage() between service worker and prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

const uid = token();

const PAGE_URL = `resources/postmessage-to-service-worker.html?uid=${uid}`;
const WORKER_URL = 'resources/postmessage-to-client-worker.js';

// Message sequence: prerendered page => service worker => prerendered page =>
// main page
promise_test(async t => {
  const registration =
    await service_worker_unregister_and_register(t, WORKER_URL, PAGE_URL);
  t.add_cleanup(() => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');

  const bc = new PrerenderChannel('test-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });
  window.open(PAGE_URL, '_blank', 'noopener');

  const result = await gotMessage;
  const expected = [
    {event: 'started waiting ServiceWorker.postMessage',  prerendering: true},
    {event: 'prerendering change',                        prerendering: false},
    {event: 'postmessage to client',                      prerendering: false},
    {event: 'finished waiting ServiceWorker.postMessage', prerendering: false},
  ];
  assert_equals(result.length, expected.length);
  for (let i = 0; i < result.length; i++) {
    assert_equals(result[i].event, expected[i].event, `event[${i}]`);
    assert_equals(result[i].prerendering, expected[i].prerendering,
      `prerendering[${i}]`);
  }

  // Send a close signal to PrerenderEventCollector on the prerendered page.
  new PrerenderChannel('close', uid).postMessage('');
}, 'ServiceWorker#postMessage() from a prerendered page should be deferred ' +
   'until page activation.');
</script>
